Removing explicit CatchUp state
